webapp/requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
WTForms==2.3.3
Werkzeug==2.2.2
Jinja2==3.1.1
MarkupSafe==2.1.1
itsdangerous==2.1.1
click==8.1.2
gunicorn==20.1.0
google-cloud-storage==2.3.0
google-cloud-bigquery==3.0.1
google-cloud-datastore